% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafa_clusters.R
\name{mafa_clusters}
\alias{mafa_clusters}
\title{Add clusters of spatial entities}
\usage{
mafa_clusters(sf, clustering, cutting)
}
\arguments{
\item{sf}{sf object.}

\item{clustering}{\code{hclust} object, e.g. the result of \code{mafa_clustering()}.}

\item{cutting}{numeric value of cutting threshold.}
}
\value{
sf object with clusters as a new column.
}
\description{
Add clusters of spatial entities from a selected cutting threshold of a \code{hclust} object.
}
\seealso{
\code{\link[=mafa_clustering]{mafa_clustering()}}
}
