% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mafa_mbboxoriented.R
\name{mafa_mbboxoriented}
\alias{mafa_mbboxoriented}
\title{Minimal oriented bounding rectangles}
\usage{
mafa_mbboxoriented(sfpolygon)
}
\arguments{
\item{sfpolygon}{sf object composed of polygons.}
}
\value{
sf object composed of minimal oriented bounding rectangles.
}
\description{
Create minimal oriented bounding rectangles of polygons.
}
