% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_circularity.R
\name{morphalr_circularity}
\alias{morphalr_circularity}
\title{Miller circularity index}
\usage{
morphalr_circularity(sfpolygons)
}
\arguments{
\item{sfpolygons}{sf object composed of polygons.}
}
\value{
\code{sfpoygons} with index result in a new column, named \code{miller_index}.
}
\description{
Compute Miller circularity index of polygons.
}
\details{
\eqn{I_M = \frac{4\pi A_p}{P_p^2}}, with \eqn{A_p} the area of a polygon and \eqn{P_{p}} its perimeter.
Miller circularity index is the ratio between the area of the polygon and a theoretical circle with the same perimeter.
In this sense, the index is also written as a similarity index to a circle \eqn{I_M = \frac{A_p}{\pi(\frac{P_p}{2\pi})^2}}.
Miller index is normalized between \verb{[0-1]}.
}
\references{
Robert, Sandrine, Éric Grosso, Pascal Chareille, et Hélène Noizet. 2014. « MorphAL (Morphological Analysis) : un outil d’analyse de morphologie urbaine ». In Archéologie de l’espace urbain, édité par Elisabeth Lorans et Xavier Rodier, 451‑63. Perspectives Villes et Territoires. Tours: Presses universitaires François-Rabelais. \url{https://doi.org/10.4000/books.pufr.7717}.
}
