% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_clustering.R
\name{morphalr_clustering}
\alias{morphalr_clustering}
\title{Compute clustering of spatial entities}
\usage{
morphalr_clustering(sf, pca_center, pca_scale, hca_method)
}
\arguments{
\item{sf}{sf object.}

\item{pca_center}{logical or numeric value, centering option as in \code{dudi.pca()}. TRUE: centering by the mean, FALSE: no centering}

\item{pca_scale}{logical value indicating whether the column vectors should be scaled as in \code{dudi.pca()}.}

\item{hca_method}{character string defining the clustering method implemented in \code{fastcluster::hclust()}: \code{"single"}, \code{"single"},
\code{"complete"}, \code{"ward.D"}, \code{"ward.D2"}, \code{"mcquitty"}, \code{"centroid"} or \code{"median"}.}
}
\value{
clustering object of class \code{hclust}.
}
\description{
Compute clustering of spatial entities based on HCA on distance matrix of coordinates of PCA.
}
\seealso{
\code{\link[=dudi.pca]{dudi.pca()}}

\code{\link[fastcluster:hclust]{fastcluster::hclust()}}
}
