% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_clusters.R
\name{morphalr_clusters}
\alias{morphalr_clusters}
\title{Add clusters of spatial entities}
\usage{
morphalr_clusters(sf, clustering, cutting)
}
\arguments{
\item{sf}{sf object.}

\item{clustering}{hclust object, e.g. the result of \code{morphalr_clustering()}.}

\item{cutting}{numeric value of cutting threshold.}
}
\value{
sf object with clusters as a new variable.
}
\description{
Add clusters of spatial entities from selected cutting threshold of the \code{morphalr_clustering()} result.
}
\seealso{
\code{\link[=morphalr_clustering]{morphalr_clustering()}}
}
