% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_clusters_mean.R
\name{morphalr_clusters_mean}
\alias{morphalr_clusters_mean}
\title{Mean of center-scale values of columns by cluster}
\usage{
morphalr_clusters_mean(sf, clustering, cutting)
}
\arguments{
\item{sf}{sf object.}

\item{clustering}{hclust object, e.g. the result of \code{morphalr_clustering()}.}

\item{cutting}{numeric value of cutting threshold.}
}
\value{
a tibble.
}
\description{
Compute mean of center-scale values of columns group by cluster.
}
\seealso{
\code{\link[=morphalr_clustering]{morphalr_clustering()}}
}
