% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_mbboxoriented.R
\name{morphalr_mbboxoriented}
\alias{morphalr_mbboxoriented}
\title{Create minimal bounding rectangles}
\usage{
morphalr_mbboxoriented(sfpolygon)
}
\arguments{
\item{sfpolygon}{A sf object composed of polygons.}
}
\value{
A sf object composed of minimal oriented bounding rectangle.
}
\description{
Create minimal bounding oriented rectangles of polygons.
}
