% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_segment_orientation.R
\name{morphalr_segment_orientation}
\alias{morphalr_segment_orientation}
\title{Segment orientation}
\usage{
morphalr_segment_orientation(sfsegments, looking = "N", perpendicular = FALSE)
}
\arguments{
\item{sfsegments}{sf object with sfc as \code{"MULTILINESTRING"} or \code{"LINESTRING"}.}

\item{looking}{look direction parameter: \code{"N"} North or \code{"E"} East (default is North).}

\item{perpendicular}{logical parameter. If \code{TRUE}, perpendiculars to orientations between \verb{[0;90]} (i.e. \verb{[-90;0]}) degrees are recalculated in the interval \verb{[0;90]} (default is \code{FALSE}).}
}
\value{
sf object composed of segments (i.e. lines) with orientations.
}
\description{
Compute orientations of segments (i.e. lines).
}
\references{
Robert, Sandrine, Éric Grosso, Pascal Chareille, et Hélène Noizet. 2014. « MorphAL (Morphological Analysis) : un outil d’analyse de morphologie urbaine ». In Archéologie de l’espace urbain, édité par Elisabeth Lorans et Xavier Rodier, 451‑63. Perspectives Villes et Territoires. Tours: Presses universitaires François-Rabelais. \url{https://doi.org/10.4000/books.pufr.7717}.
}
