% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_utils.R
\name{morphalr_utils_mbboxoriented}
\alias{morphalr_utils_mbboxoriented}
\title{Create minBBox of a polygon}
\usage{
morphalr_utils_mbboxoriented(sfpolygon)
}
\arguments{
\item{sfpolygon}{A sf object composed of polygon}
}
\value{
A sf object composed of minimal bounding box oriented rectangle
}
\description{
Create minimal bounding oriented box rectangles of a polygon
}
