% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphal_geom_to_segment.R
\name{morphal_geom_to_segment}
\alias{morphal_geom_to_segment}
\title{Extract segments of polygons or linestrings.}
\usage{
morphal_geom_to_segment(sfobject, to = "MULTILINESTRING")
}
\arguments{
\item{sfobject}{A sf object (polygons or linestring)}

\item{to}{Type of sf output (MULTILINESTRING or LINESTRING). Default is MULTILINESTRING.}
}
\value{
A sf object composed of segments of polygons or linestrings
}
\description{
Extraction of segments (lines) composing polygons or linestrings.
}
\author{
Diego HH
}
