% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphal_segment_orientation.R
\name{morphal_segment_orientation}
\alias{morphal_segment_orientation}
\title{Compute segment orientation.}
\usage{
morphal_segment_orientation(sfsegments)
}
\arguments{
\item{sfsegments}{A sf object}
}
\value{
A sf object composed of lines with orientations
}
\description{
Compute orientations of segments (lines).
}
