% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_dsc.R
\name{morphalr_dsc}
\alias{morphalr_dsc}
\title{Compute distance to convex hull}
\usage{
morphalr_dsc(sfpolygons)
}
\arguments{
\item{sfpolygons}{A sf object composed of polygons}
}
\value{
\code{sfpoygons} with DSC as a new variable
}
\description{
Compute the distance of a polygon to its convex hull (DSC).
}
\details{
\eqn{DSC = 1 - \frac{A_p}{A_{ch}}}, with \eqn{A_p} the area of a polygon and \eqn{A_{ch}} the area of its convex hull.
DSC is normalized between \verb{[0-1]}.
}
\references{
Robert, Sandrine, Éric Grosso, Pascal Chareille, et Hélène Noizet. 2014. « MorphAL (Morphological Analysis) : un outil d’analyse de morphologie urbaine ». In Archéologie de l’espace urbain, édité par Elisabeth Lorans et Xavier Rodier, 451‑63. Perspectives Villes et Territoires. Tours: Presses universitaires François-Rabelais. \url{https://doi.org/10.4000/books.pufr.7717}.
}
