% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_dsr.R
\name{morphalr_dsr}
\alias{morphalr_dsr}
\title{Distance to minimal bounding rectangle}
\usage{
morphalr_dsr(sfpolygons)
}
\arguments{
\item{sfpolygons}{sf object composed of polygons.}
}
\value{
\code{sfpoygons} with DSR as a new column.
}
\description{
Compute the distance of a polygon to its minimal bounding rectangle (DSR).
}
\details{
\eqn{DSR = 1 - \frac{A_p}{A_{mbb}}}, with \eqn{A_p} the area of a polygon and \eqn{A_{mbb}}
the area of its minimal oriented bounding box (rectangle). DSR is normalized between \verb{[0-1]}.
}
\references{
Ali, Atef Bel Hadj. 2001. « Qualité géométrique des entités géographiques surfaciques. Application à l’appariement et définition d’une typologie des écarts geométriques ». Phdthesis, université Gustave Effeil ; Anciennement Université de Marne La vallée. \url{https://hal.science/tel-03244834}, p. 92.

Robert, Sandrine, Éric Grosso, Pascal Chareille, et Hélène Noizet. 2014. « MorphAL (Morphological Analysis) : un outil d’analyse de morphologie urbaine ». In Archéologie de l’espace urbain, édité par Elisabeth Lorans et Xavier Rodier, 451‑63. Perspectives Villes et Territoires. Tours: Presses universitaires François-Rabelais. \url{https://doi.org/10.4000/books.pufr.7717}.
}
