% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphalr_mbboxoriented.R
\name{morphalr_mbboxoriented}
\alias{morphalr_mbboxoriented}
\title{Minimal oriented bounding rectangles}
\usage{
morphalr_mbboxoriented(sfpolygon)
}
\arguments{
\item{sfpolygon}{sf object composed of polygons.}
}
\value{
sf object composed of minimal oriented bounding rectangles.
}
\description{
Create minimal oriented bounding rectangles of polygons.
}
