% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rouen_1827}
\alias{rouen_1827}
\title{Rouen in 1827}
\format{
A simple feature with 21,975 features and 1 field:
\describe{
\item{rowid}{identifiers}
\item{geometry}{geometry of polygons}
}
}
\source{
Lise Levieux, Guillaume Gaillard, 2022, ODbL Licence, \url{http://www.fabriquenumeriquedupasse.fr/explore/dataset/rouen_parcelles_1827/information/}
}
\usage{
rouen_1827
}
\description{
Parcels of the city of Rouen in 1827
}
\keyword{datasets}
